Move helper function to own class
This is needed so that we can reuse the same code when checking custom
roles.
